backend/requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
python-dotenv>=0.19.2
pymongo>=4.6.1
dnspython>=2.4.2